Created a website based on animation,transform and transition
Fixed navbar animated on scroll. A button animated with particles coming out from underneath it when clicking on it, plus another button/link animated with a hover state that a line goes around it, Animated Hamburger Icon on Click  using transform and transition + used clip-path: circle when opening a hamburger menu on small screen.