Started the Final Project- Luxurious Fine-Jewelry e commerce website
Created a Logo in illustrator,Navbar that fits the name (color and style:slides). Decided to have the landing page gets covered by the carousel that covers the whole page. Created beautiful backgrounds using photoshop and illustrator to present the three main recent collections. Added some detail about each collection and put a button for users if they want to see the collection.