Fixed the cart for any screen size, created a contact form in Contact Us page
I made the cart icon in illustrator in black and white colors suitable for different screensizes and backgrounds for more contrast. I implemented them in two spots and used display prop to show and hide them when user clicks on purchase button, also they can see how many items is in the cart.  Created a contact form in Contact Us page and worked on its functionality.